Move tests to RtacResourceImplTest
